# Swagger/redis-api.yaml
swagger: '2.0'
info:
  title: First Draft
  version: "1.0.0"
# the domain of the service
host: firstdraft.com
...
schemes:
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /test-plan/complete/:
    get:
      summary: Get all tests that are available to the testee
      operationId: testPlan.complete
      tags:
        - Test Plan
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/AvailableTests'
        404:
          description: Not Found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  AvailableTests:
    type: array
    items:
      $ref: '#/definitions/Test'
...
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
      message:
        type: string
...
  Test:
    type: object
    properties:
      id:
        type: integer
        description: Record Id
      created:
        type: integer
        format: dateTime
      authorId:
        type: integer
        description: Id of the user that created the test.
      testLevelId:
        type: integer
        description: Test level number (1 min, 10 max)
      sourceId:
        type: integer
        description: Id of the user being evaluated
      targetId:
        type: integer
        description: Source language id
      minCorrect:
        type: integer
        description: Minimum correct in order to pass
      maxRetries:
        type: integer
        description: Max times that a user can retry this test
      description:
        type: string
...